# backend/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for ThulirFarm dairy management application
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface SendOtpResponse {
  success: boolean;
  message: string;
  /** Demo OTP (remove in production) */
  demoOtp: string;
  /** Expiry time in seconds */
  expiresIn: number;
}

export interface VerifyOtpResponse {
  success: boolean;
  verified: boolean;
}

export interface DiagnoseRequest {
  symptoms: string[];
  customNote?: string;
  animalName?: string;
  animalType?: string;
}

export type DiagnoseResponseRiskLevel =
  (typeof DiagnoseResponseRiskLevel)[keyof typeof DiagnoseResponseRiskLevel];

export const DiagnoseResponseRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface DiagnoseResponse {
  summary?: string;
  summaryTamil?: string;
  possibleCauses?: string[];
  immediateActions?: string[];
  riskLevel?: DiagnoseResponseRiskLevel;
  medicine?: string | null;
  tamilAdvice?: string;
  homeRemedy?: string | null;
  nextSteps?: string;
  callVetImmediately?: boolean;
}

export type VoiceCommandRequestAnimalsItem = {
  id?: string;
  name?: string;
  type?: string;
};

export interface VoiceCommandRequest {
  transcript: string;
  animals?: VoiceCommandRequestAnimalsItem[];
}

export type VoiceCommandResponseAction =
  (typeof VoiceCommandResponseAction)[keyof typeof VoiceCommandResponseAction];

export const VoiceCommandResponseAction = {
  log_milk: "log_milk",
  report_problem: "report_problem",
  add_expense: "add_expense",
  navigate: "navigate",
  unknown: "unknown",
} as const;

export type VoiceCommandResponseParams = { [key: string]: unknown };

export interface VoiceCommandResponse {
  action?: VoiceCommandResponseAction;
  /**
   * @minimum 0
   * @maximum 1
   */
  confidence?: number;
  params?: VoiceCommandResponseParams;
  confirmationText?: string;
  confirmationTamil?: string;
}

export interface TranscribeResponse {
  transcript: string;
}

export type ChatRequestHistoryItemRole =
  (typeof ChatRequestHistoryItemRole)[keyof typeof ChatRequestHistoryItemRole];

export const ChatRequestHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type ChatRequestHistoryItem = {
  role?: ChatRequestHistoryItemRole;
  content?: string;
};

export type ChatRequestLanguage =
  (typeof ChatRequestLanguage)[keyof typeof ChatRequestLanguage];

export const ChatRequestLanguage = {
  ta: "ta",
  te: "te",
  kn: "kn",
  ml: "ml",
  hi: "hi",
  en: "en",
} as const;

export interface ChatRequest {
  message: string;
  history?: ChatRequestHistoryItem[];
  language?: ChatRequestLanguage;
}

export interface ChatResponse {
  response: string;
}

export type RationRequestLanguage =
  (typeof RationRequestLanguage)[keyof typeof RationRequestLanguage];

export const RationRequestLanguage = {
  ta: "ta",
  te: "te",
  kn: "kn",
  ml: "ml",
  hi: "hi",
  en: "en",
} as const;

export interface RationRequest {
  animalType: string;
  breed: string;
  weightKg: number;
  milkProductionL: number;
  language?: RationRequestLanguage;
}

export type RationResponseGreenFodder = {
  quantity?: string;
  examples?: string;
};

export type RationResponseDryFodder = {
  quantity?: string;
  examples?: string;
};

export type RationResponseConcentrate = {
  quantity?: string;
  composition?: string;
};

export interface RationResponse {
  summary?: string;
  summaryLocal?: string;
  greenFodder?: RationResponseGreenFodder;
  dryFodder?: RationResponseDryFodder;
  concentrate?: RationResponseConcentrate;
  mineralMix?: string;
  water?: string;
  totalCost?: string;
  tips?: string[];
}

export type SendOtpBody = {
  /**
   * 10-digit Indian phone number
   * @pattern ^[6-9]\d{9}$
   */
  phone: string;
};

export type VerifyOtpBody = {
  phone: string;
  otp: string;
};

export type TranscribeAudioBody = {
  audio?: Blob;
};