# backend/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for ThulirFarm dairy management application
 * OpenAPI spec version: 0.1.0
 */
/// <reference lib="dom" />
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Send OTP to phone number for authentication
 * @summary Send OTP
 */
export const sendOtpBodyPhoneRegExp = new RegExp("^[6-9]\\d{9}$");

export const SendOtpBody = zod.object({
  phone: zod
    .string()
    .regex(sendOtpBodyPhoneRegExp)
    .describe("10-digit Indian phone number"),
});

export const SendOtpResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  demoOtp: zod.string().describe("Demo OTP (remove in production)"),
  expiresIn: zod.number().describe("Expiry time in seconds"),
});

/**
 * Verify OTP for authentication
 * @summary Verify OTP
 */
export const VerifyOtpBody = zod.object({
  phone: zod.string(),
  otp: zod.string(),
});

export const VerifyOtpResponse = zod.object({
  success: zod.boolean(),
  verified: zod.boolean(),
});

/**
 * Get AI-powered veterinary diagnosis from symptoms
 * @summary AI Veterinary Diagnosis
 */
export const DiagnoseBody = zod.object({
  symptoms: zod.array(zod.string()),
  customNote: zod.string().optional(),
  animalName: zod.string().optional(),
  animalType: zod.string().optional(),
});

export const DiagnoseResponse = zod.object({
  summary: zod.string().optional(),
  summaryTamil: zod.string().optional(),
  possibleCauses: zod.array(zod.string()).optional(),
  immediateActions: zod.array(zod.string()).optional(),
  riskLevel: zod.enum(["low", "medium", "high", "critical"]).optional(),
  medicine: zod.string().nullish(),
  tamilAdvice: zod.string().optional(),
  homeRemedy: zod.string().nullish(),
  nextSteps: zod.string().optional(),
  callVetImmediately: zod.boolean().optional(),
});

/**
 * Parse voice command transcript into structured action
 * @summary Parse Voice Command
 */
export const ParseVoiceCommandBody = zod.object({
  transcript: zod.string(),
  animals: zod
    .array(
      zod.object({
        id: zod.string().optional(),
        name: zod.string().optional(),
        type: zod.string().optional(),
      }),
    )
    .optional(),
});

export const parseVoiceCommandResponseConfidenceMin = 0;
export const parseVoiceCommandResponseConfidenceMax = 1;

export const ParseVoiceCommandResponse = zod.object({
  action: zod
    .enum(["log_milk", "report_problem", "add_expense", "navigate", "unknown"])
    .optional(),
  confidence: zod
    .number()
    .min(parseVoiceCommandResponseConfidenceMin)
    .max(parseVoiceCommandResponseConfidenceMax)
    .optional(),
  params: zod.object({}).passthrough().optional(),
  confirmationText: zod.string().optional(),
  confirmationTamil: zod.string().optional(),
});

/**
 * Transcribe audio using Whisper
 * @summary Transcribe Audio
 */
export const TranscribeAudioBody = zod.object({
  audio: zod.instanceof(File).optional(),
});

export const TranscribeAudioResponse = zod.object({
  transcript: zod.string(),
});

/**
 * Chat with AI dairy expert assistant
 * @summary Chat with GauGuru
 */
export const ChatWithGauGuruBody = zod.object({
  message: zod.string(),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]).optional(),
        content: zod.string().optional(),
      }),
    )
    .optional(),
  language: zod.enum(["ta", "te", "kn", "ml", "hi", "en"]).optional(),
});

export const ChatWithGauGuruResponse = zod.object({
  response: zod.string(),
});

/**
 * Calculate daily feed ration for animal
 * @summary Calculate Feed Ration
 */
export const CalculateRationBody = zod.object({
  animalType: zod.string(),
  breed: zod.string(),
  weightKg: zod.number(),
  milkProductionL: zod.number(),
  language: zod.enum(["ta", "te", "kn", "ml", "hi", "en"]).optional(),
});

export const CalculateRationResponse = zod.object({
  summary: zod.string().optional(),
  summaryLocal: zod.string().optional(),
  greenFodder: zod
    .object({
      quantity: zod.string().optional(),
      examples: zod.string().optional(),
    })
    .optional(),
  dryFodder: zod
    .object({
      quantity: zod.string().optional(),
      examples: zod.string().optional(),
    })
    .optional(),
  concentrate: zod
    .object({
      quantity: zod.string().optional(),
      composition: zod.string().optional(),
    })
    .optional(),
  mineralMix: zod.string().optional(),
  water: zod.string().optional(),
  totalCost: zod.string().optional(),
  tips: zod.array(zod.string()).optional(),
});