# backend/lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for ThulirFarm dairy management application
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  response: string;
}