# backend/lib/api-zod/src/generated/types/diagnoseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for ThulirFarm dairy management application
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnoseResponseRiskLevel } from "./diagnoseResponseRiskLevel";

export interface DiagnoseResponse {
  summary?: string;
  summaryTamil?: string;
  possibleCauses?: string[];
  immediateActions?: string[];
  riskLevel?: DiagnoseResponseRiskLevel;
  medicine?: string | null;
  tamilAdvice?: string;
  homeRemedy?: string | null;
  nextSteps?: string;
  callVetImmediately?: boolean;
}