# backend/lib/api-zod/src/generated/types/voiceCommandResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for ThulirFarm dairy management application
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceCommandResponseAction } from "./voiceCommandResponseAction";
import type { VoiceCommandResponseParams } from "./voiceCommandResponseParams";

export interface VoiceCommandResponse {
  action?: VoiceCommandResponseAction;
  /**
   * @minimum 0
   * @maximum 1
   */
  confidence?: number;
  params?: VoiceCommandResponseParams;
  confirmationText?: string;
  confirmationTamil?: string;
}